Updated timer reminder type.

+++ app/src/main/java/com/cray/software/justreminder/ArchiveActivity.java
@@ -529,7 +529,7 @@ public class ArchiveActivity extends AppCompatActivity
                 int month = c.getInt(c.getColumnIndex(Constants.COLUMN_MONTH));
                 int year = c.getInt(c.getColumnIndex(Constants.COLUMN_YEAR));
                 int repCode = c.getInt(c.getColumnIndex(Constants.COLUMN_REPEAT));
-                int repTime = c.getInt(c.getColumnIndex(Constants.COLUMN_REMIND_TIME));
+                long repTime = c.getLong(c.getColumnIndex(Constants.COLUMN_REMIND_TIME));
                 int isDone = c.getInt(c.getColumnIndex(Constants.COLUMN_IS_DONE));
                 double lat = c.getDouble(c.getColumnIndex(Constants.COLUMN_LATITUDE));
                 double lon = c.getDouble(c.getColumnIndex(Constants.COLUMN_LONGITUDE));
